dana-page-builder/requirements.txt
# Dana Page Builder - Python Dependencies

# Google APIs
gspread==5.12.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Image Processing
Pillow==10.1.0
numpy==1.26.2

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Template Engine
Jinja2==3.1.2

# Utilities
python-dotenv==1.0.0
